coordinator: move Anchoring events to common interfaces packages
