+++ coordinator/ethereum/message-anchoring-i2/src/main/kotlin/linea/anchoring/MessageAnchoringService.kt
@@ -1,8 +1,8 @@
 package linea.anchoring
 
 import io.vertx.core.Vertx
-import linea.anchoring.events.L1RollingHashUpdatedEvent
-import linea.anchoring.events.MessageSentEvent
+import linea.contract.events.L1RollingHashUpdatedEvent
+import linea.contract.events.MessageSentEvent
 import linea.contract.l2.L2MessageServiceSmartContractClient
 import linea.domain.BlockParameter
 import linea.domain.CommonDomainFunctions
