+++ jvm-libs/linea/clients/interfaces/src/main/kotlin/linea/contract/events/L2RollingHashUpdatedEvent.kt
@@ -1,4 +1,4 @@
-package linea.anchoring.events
+package linea.contract.events
 
 import linea.domain.EthLog
 import linea.domain.EthLogEvent
