+++ jvm-libs/linea/clients/interfaces/src/test/kotlin/linea/contract/events/L2RollingHashUpdatedEventTest.kt
@@ -1,10 +1,10 @@
-package linea.anchoring.events
+package linea.contract.events
 
 import linea.domain.EthLog
 import linea.domain.EthLogEvent
 import linea.kotlin.decodeHex
 import linea.kotlin.toULongFromHex
-import org.assertj.core.api.Assertions.assertThat
+import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 
 class L2RollingHashUpdatedEventTest {
@@ -40,6 +40,6 @@ class L2RollingHashUpdatedEventTest {
       log = ethLog
     )
 
-    assertThat(result).isEqualTo(expectedEthLogEvent)
+    Assertions.assertThat(result).isEqualTo(expectedEthLogEvent)
   }
 }
